--- src/latexify/math_symbols.py
@@ -1,55 +1,0 @@
-"""Utilities to manipulate math symbols."""
-
-from __future__ import annotations
-
-_MATH_SYMBOLS = {
-    "aleph",
-    "alpha",
-    "beta",
-    "beth",
-    "chi",
-    "daleth",
-    "delta",
-    "digamma",
-    "epsilon",
-    "eta",
-    "gamma",
-    "gimel",
-    "hbar",
-    "infty",
-    "iota",
-    "kappa",
-    "lambda",
-    "mu",
-    "nabla",
-    "nu",
-    "omega",
-    "phi",
-    "pi",
-    "psi",
-    "rho",
-    "sigma",
-    "tau",
-    "theta",
-    "upsilon",
-    "varepsilon",
-    "varkappa",
-    "varphi",
-    "varpi",
-    "varrho",
-    "varsigma",
-    "vartheta",
-    "xi",
-    "zeta",
-    "Delta",
-    "Gamma",
-    "Lambda",
-    "Omega",
-    "Phi",
-    "Pi",
-    "Psi",
-    "Sigma",
-    "Theta",
-    "Upsilon",
-    "Xi",
-}
